Move modlist to json file
This is currently the bare minimum implmentation. There is no error handling, and any error
will currently lead to a panic (mising file, failure to read, failure to parse, etc.)

Still left to do is to flesh out this data format and handle auto-updating of the file.
(And Error Handling)