Working custom section
Some size offsets are still hardcoded and thus won't work for all sections, or for
adding more than one section.

The serialization process likely needs to dynamically calculte the relevant
pointers based on the data being serialized. One possible way to accomplish this
would be to use a 'raw' version of the XBE structure and a different version
that's easier to dynically edit then create a raw structure from.